voice-to-action/python_app/requirements.txt
PySide6>=6.6.0
requests>=2.31.0
python-dotenv>=1.0.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-ApplicationServices>=10.0
pyobjc-framework-Quartz>=10.0
numpy>=1.24.0
sounddevice>=0.4.6
websockets>=11.0.0
pygame>=2.6.1